requirements.txt
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-dotenv==1.0.1
clerk-backend-api
fastapi
httpx
uvicorn[standard]
